# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "WORLD EXPLORER API",
    "description": "API for managing a collection of popular destinations all over the world.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/destinations/": {
      
      
      "get": {
        "summary": "Get all destinations",
        "description": "Retrieve a list of all destinations stored in the database.",
        "responses": {
          "200": {
            "description": "List of destinations"
          }
        }
      },
      
      
      "post": {
        "summary": "Add a new destination",
        "description": "Create and store a new fictional destination in the database.",
        "parameters": [
          {
            "in": "body",
            "name": "destination",
            "description": "Destination object that needs to be added to the database",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Europe"
                },
                "country": {
                  "description": "string",
                  "example": "Switzerland"
                },
                "description": {
                  "description": "string",
                  "example": "Second smallest of the world continents."
                },
                "bestTimeToVisit": {
                  "description": "string",
                  "example": "April - May, July - August,"
                },
                "topAttraction": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": ["mountains", "ocean", "natural beauty"]
                },
                "source": {
                  "type": "string",
                  "example": "Italy"
                }
              },
              "required": ["name", "country", "description", "bestTimeToVisit", "topAttraction", "source"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Destination successfully created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },

 "/api/destinations/{id}": {
      "put": {
        "summary": "Update a destination",
        "description": "Update details of an existing destination by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of destination to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "destination",
            "description": "Updated destination object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "country": { "type": "string" },
                "description": { "type": "string" },
                "bestTimeToVisit": { "type": "string" },
                "topAttraction": {
                  "type": "array",
                  "items": { "type": "string" }
                },
                "source": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Destination successfully updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Destination not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a destination",
        "description": "Remove a destination from the database by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of destination to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Destination successfully deleted"
          },
          "404": {
            "description": "Destination not found"
          }
        }
      }
    }
  }
}